# dispatch-loadbalancer-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0",
    "description": "Spring Boot backend service that optimally assigns delivery orders to vehicles.\nSupports Haversine distance calculations, priority-based order handling (HIGH > MEDIUM > LOW), and vehicle capacity constraints."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/dispatch/vehicles": {
      "post": {
        "tags": [
          "dispatch-controller"
        ],
        "operationId": "takeVehicles",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOListVehicle"
                }
              }
            }
          }
        },
        "description": "Accepts and stores a list of delivery vehicles. Vehicles include live location and capacity data used in dispatch planning."
      }
    },
    "/api/dispatch/orders": {
      "post": {
        "tags": [
          "dispatch-controller"
        ],
        "operationId": "takeOrders",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOListOrder"
                }
              }
            }
          }
        },
        "description": "Accepts and stores a list of delivery orders including coordinates, weight, and priority."
      }
    },
    "/api/dispatch/plan": {
      "get": {
        "tags": [
          "dispatch-controller"
        ],
        "operationId": "getPlans",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTOListDispatchPlan"
                }
              }
            }
          }
        },
        "description": "Returns an optimized dispatch plan per vehicle based on the greedy distance + priority algorithm."
      }
    },
    "/api/dispatch/fullPlan": {
      "get": {
        "tags": [
          "dispatch-controller"
        ],
        "operationId": "getFullPlans",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDTODispatchResult"
                }
              }
            }
          }
        },
        "description": "Returns a complete dispatch summary with all plan metrics and a list of unassignable orders."
      }
    }
  },
  "components": {
    "schemas": {
      "VehicleListRequest": {
        "type": "object",
        "properties": {
          "vehicles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehicleRequest"
            },
            "minItems": 1
          }
        },
        "required": [
          "vehicles"
        ]
      },
      "VehicleRequest": {
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string",
            "minLength": 1
          },
          "capacity": {
            "type": "number",
            "format": "double",
            "minimum": 1
          },
          "currentLatitude": {
            "type": "number",
            "format": "double"
          },
          "currentLongitude": {
            "type": "number",
            "format": "double"
          },
          "currentAddress": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "currentAddress",
          "currentLatitude",
          "currentLongitude",
          "vehicleId"
        ]
      },
      "ResponseDTOListVehicle": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vehicle"
            }
          }
        }
      },
      "Vehicle": {
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "currentLatitude": {
            "type": "number",
            "format": "double"
          },
          "currentLongitude": {
            "type": "number",
            "format": "double"
          },
          "currentAddress": {
            "type": "string"
          }
        },
        "description": "Represents a delivery vehicle and its live position."
      },
      "OrderListRequest": {
        "type": "object",
        "properties": {
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderRequest"
            },
            "minItems": 1
          }
        },
        "required": [
          "orders"
        ]
      },
      "OrderRequest": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "minLength": 1
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "maximum": 90,
            "minimum": -90
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "maximum": 180,
            "minimum": -180
          },
          "address": {
            "type": "string",
            "minLength": 1
          },
          "packageWeight": {
            "type": "number",
            "format": "double",
            "minimum": 1
          },
          "priority": {
            "type": "string",
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW"
            ]
          }
        },
        "required": [
          "address",
          "latitude",
          "longitude",
          "orderId",
          "priority"
        ]
      },
      "Order": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "address": {
            "type": "string"
          },
          "packageWeight": {
            "type": "integer",
            "format": "int32"
          },
          "priority": {
            "type": "string",
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW"
            ]
          }
        },
        "description": "Represents a delivery order with geolocation and priority."
      },
      "ResponseDTOListOrder": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          }
        }
      },
      "DispatchPlan": {
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string"
          },
          "totalLoad": {
            "type": "integer",
            "format": "int32"
          },
          "totalDistance": {
            "type": "string"
          },
          "assignedOrders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          }
        },
        "description": "Plan assigned to a single vehicle, including all orders, total load, and distance travelled."
      },
      "ResponseDTOListDispatchPlan": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DispatchPlan"
            }
          }
        }
      },
      "DispatchResult": {
        "type": "object",
        "properties": {
          "totalOrders": {
            "type": "integer",
            "format": "int32"
          },
          "averageDistance": {
            "type": "string"
          },
          "totalDistance": {
            "type": "string"
          },
          "totalLoad": {
            "type": "integer",
            "format": "int32"
          },
          "dispatchPlan": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DispatchPlan"
            }
          },
          "unassignedOrders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          }
        },
        "description": "Aggregated result of dispatching process: includes metrics like total orders, load, average and total distance, vehicle plans, and unassigned orders."

      },
      "ResponseDTODispatchResult": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/DispatchResult"
          }
        }
      }
    }
  }
}